+++ tests/test_visual_and_terminology_quality.py
@@ -28,6 +28,7 @@ FIGURES = (
     "p22_simultaneous_refinement_chain_certification.svg",
     "p23_adaptive_descriptor_selection_certification.svg",
     "p24_anytime_adaptive_refinement_certification.svg",
+    "p25_directed_influence_scale_certification.svg",
     "universal_proof_ladder.svg",
     "theory_comparison_map.svg",
 )